[CBRD-23109] HA server state header (#1759)
http://jira.cubrid.org/browse/CBRD-23109

Move HA server state from boot.h to its own header.
